Mission Calculator, First Version
First version of a program which, given certain information such as the distance of an orbiting object from the sun, as well as the angular separation between Earth and the object on a given date, calculates the mission details for a mission which would travel on a Hohmann orbit from Earth orbit to the orbiting object. 
This version assumes Earth and the orbiting object's orbits are sufficiently circular, and that the object is already in Earth's orbit.